Further Improved Execution Time.
Replaced as many instances of MeasureText() as I could with a single
call in the constructor.

As a monospace font is being used, all calls to MeasureText() in
MemoryDisplay_Paint could be removed, reducing execution time to 3.1ms.
Also removed a number of calls in DisassemblyDisplay_Paint to strings of
identical sizes.

I also moved font creation to the constructors, as the same font is used
in multiple methods.